requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
aiosqlite>=0.20.0
python-telegram-bot>=20.7
playwright>=1.40.0
